refactor: add mqtt module
Moves `parser_json` function to a new `mqtt` module.
